bitbake: toaster: Added feat to import eventlogs files as builds in toaster
New feature added to recreate command line builds that are made when toaster is off.
The feature contains a new button on the base template to access a new template.
A new model was added to register the information on the builds and generate access links
A new form was added to include the option to load specific files
This feature uses the value from the variable BB_DEFAULT_EVENTLOG to read the files created by bitbake

(Bitbake rev: b76cd36e115eee742a1ee57373856d0081127d34)

Signed-off-by: Marlon Rodriguez Garcia <marlon.rodriguez-garcia@savoirfairelinux.com>
Signed-off-by: Richard Purdie <richard.purdie@linuxfoundation.org>